Maintenance and fsck

+++ src/NyxFileSystemElementsMapping.rb
@@ -10,6 +10,8 @@ class NyxFileSystemElementsMapping
 
     # NyxFileSystemElementsMapping::register(objectuuid, elementname, location)
     def self.register(objectuuid, elementname, location)
+        raise "[error] 3127b377-7819-4b37-916c-b68cdf34264b" if location.nil?
+        raise "[error] 599f17a1-df7e-47a8-ad71-35a7f7c7c5e1" if !File.exists?(location) 
         db = SQLite3::Database.new(NyxFileSystemElementsMapping::databaseFilepath())
         db.execute "delete from mapping where _objectuuid_=?", [objectuuid]
         db.execute "insert into mapping (_objectuuid_, _name_, _location_) values ( ?, ?, ? )", [objectuuid, elementname, location]
@@ -59,9 +61,9 @@ class NyxFileSystemElementsMapping
                 next
             end
             if !File.exists?(record["location"]) then
-                puts "Incorrect record:"
+                puts "[NyxFileSystemElementsMapping] incorrect record:"
                 puts JSON.pretty_generate(record)
-                exit
+                raise "[error] 40e2306c-e892-44e9-856a-21e650b4751b"
             end
         }
     end
